+++ ntex-service/src/lib.rs
@@ -23,7 +23,9 @@ mod transform_err;
 
 pub use self::apply::{apply_fn, apply_fn_factory};
 pub use self::apply_cfg::{apply_cfg, apply_cfg_factory};
-pub use self::fn_service::{fn_factory, fn_factory_with_config, fn_service};
+pub use self::fn_service::{
+    fn_factory, fn_factory_with_config, fn_mut_service, fn_service,
+};
 pub use self::map_config::{map_config, unit_config};
 pub use self::pipeline::{pipeline, pipeline_factory, Pipeline, PipelineFactory};
 pub use self::transform::{apply, Transform};
@@ -377,7 +379,7 @@ where
 pub mod dev {
     pub use crate::apply::{Apply, ApplyServiceFactory};
     pub use crate::fn_service::{
-        FnService, FnServiceConfig, FnServiceFactory, FnServiceNoConfig,
+        FnMutService, FnService, FnServiceConfig, FnServiceFactory, FnServiceNoConfig,
     };
     pub use crate::map::{Map, MapServiceFactory};
     pub use crate::map_config::{MapConfig, UnitConfig};